msglist [nfc]: Pull out selector getBackgroundData
